EXPERIMENTAL: Committing because this is something I started this last weekend and I will not hvae much time to continue the rest of this week : ( A lot has been changed this commit so I'm not 100% sure what all has been changed, but that's what SVN is for ;) The main point of this commit is trying to get FAGs / attributes to work, some values seem to work others seem to have weird BS values :( Right now there are 3 different methods I have been given / found for obtaining the attributes, so later one someone will have to test them out and see which one is fastest, or actually works ;p

+++ D3 Adventures/Structures/SNOTable.cs
@@ -15,7 +15,7 @@ namespace D3_Adventures.Structures
         internal SNO.ClientSNOTable clientSNOTable_0;
         internal readonly Dictionary<int, SNORecord> dictionary_0;
 
-        internal SNOTable(IntPtr ptr, SNO.ClientSNOTable type)
+        internal SNOTable(uint ptr, SNO.ClientSNOTable type)
             : base(ptr)
         {
             this.dictionary_0 = new Dictionary<int, SNORecord>();
